+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Add [Volta](https://volta.sh) support @nzambello
 - Various minor `Makefile` cleanup @rpatterson
+- Improve error handling in UniversalLink @nzambello
 
 ## 13.1.2 (2021-05-26)